[node] PDO default initialization
Creates ability to specify PDO default config in device config.

This leads to a requirement for runtime default initialization (because
PDO IDs can be a function of the node ID), which leads to a requirement
to implement callbacks for loading persisted object values, so the
way that callbacks are handled is also refactored.

Callbacks are no longer required to be static, but instead Node has an
generic lifetime.